[feat] support fp8_e4m3 as a block-scale dtype
Adds the narrow scale dtype NVFP4 pairs with a per-tensor global scale, ahead
of the global scale itself.

A scale in e4m3 must be clamped into [2**-9, 448] at both ends before the cast.
Below the floor it rounds to zero and divides its operand by zero. Above the
ceiling the cast is device-dependent: CUDA saturates 1e10 to 448, but CPU
returns NaN, e4m3's top encoding, so an unclamped overflow would poison a whole
tensor's scales on CPU alone. Both bounds come from a _STR_TO_QMIN table beside
_STR_TO_QMAX, read through str_to_qmin.

The cast rounds up rather than to nearest, as the e8m0 branch's ceil already
does, so a scale always covers its block's amax and never leaves the block's
largest element to be clipped. That costs at most one representable step:
measured worst-case utilisation 0.967 of qmax.

Triton's scaled kernels cast their scale loads to fp32, a no-op for fp32 scales
so existing configs compile unchanged.

Tests assert the clamp across CPU and CUDA, since a CUDA-only assertion passes
whether or not the ceiling is there. The precision oracle mirrors e4m3's
rounding, deriving the ceiling independently by lookup over the format's
representable grid. Its utilisation bound now exempts floored blocks: fp8_e5m2's
qmax of 57344 puts every scale three decades under e4m3's floor, which is what
the global scale exists to repair.

+++ src/kernel/backends/triton/gemm.py
@@ -333,16 +333,17 @@ def _scaled_mm_kernel(
             a = tl.load(a_ptrs, mask=m_mask[:, None] & k_mask[None, :], other=0.0)
             b = tl.load(b_ptrs, mask=k_mask[:, None] & n_mask[None, :], other=0.0)
             block_acc += tl.dot(a, b).to(tl.float32)
+        # Cast so a narrow (e4m3) scale becomes arithmetic-capable; a no-op for fp32.
         sa = tl.load(
             sa_ptr + offs_m * stride_sam + scale_block_idx * stride_sak,
             mask=m_mask,
             other=0.0,
-        )
+        ).to(tl.float32)
         sb = tl.load(
             sb_ptr + scale_block_idx * stride_sbk + offs_n * stride_sbn,
             mask=n_mask,
             other=0.0,
-        )
+        ).to(tl.float32)
         acc += sa[:, None] * block_acc * sb[None, :]
     # Bias is unscaled: apply it to acc, not each scaled block partial.
     if HAS_BIAS:
@@ -799,6 +800,7 @@ def _scaled_grouped_mm_kernel(
                         b_ptrs, mask=k_mask[:, None] & n_mask[None, :], other=0.0
                     )
                     block_acc += tl.dot(a, b).to(tl.float32)
+                # Cast so a narrow (e4m3) scale is arithmetic-capable; fp32 is a no-op.
                 sa = tl.load(
                     sa_ptr
                     + (0 if A_IS_2D else g * stride_sag)
@@ -806,7 +808,7 @@ def _scaled_grouped_mm_kernel(
                     + scale_block_idx * stride_sak,
                     mask=m_mask,
                     other=0.0,
-                )
+                ).to(tl.float32)
                 sb = tl.load(
                     sb_ptr
                     + (0 if B_IS_2D else g * stride_sbg)
@@ -814,7 +816,7 @@ def _scaled_grouped_mm_kernel(
                     + (n_start + offs_n) * stride_sbn,
                     mask=n_mask,
                     other=0.0,
-                )
+                ).to(tl.float32)
                 acc += sa[:, None] * block_acc * sb[None, :]
             # Bias is unscaled: apply it to acc, not each scaled block partial.
             if HAS_BIAS:
